Remove deprecated liblms and libxmss
