+++ CMakeLists.txt
@@ -700,10 +700,8 @@ add_option(WOLFSSL_LMSSHA256192
 
 if (WOLFSSL_LMS)
     list(APPEND WOLFSSL_DEFINITIONS "-DWOLFSSL_HAVE_LMS")
-    list(APPEND WOLFSSL_DEFINITIONS "-DWOLFSSL_WC_LMS")
 
     set_wolfssl_definitions("WOLFSSL_HAVE_LMS" RESULT)
-    set_wolfssl_definitions("WOLFSSL_WC_LMS"   RESULT)
 
     if (WOLFSSL_LMSSHA256192)
         list(APPEND WOLFSSL_DEFINITIONS "-DWOLFSSL_LMS_SHA256_192")
