+++ wolfcrypt/benchmark/benchmark.c
@@ -178,20 +178,10 @@
     #endif
 #endif
 #if defined(WOLFSSL_HAVE_LMS) && !defined(WOLFSSL_LMS_VERIFY_ONLY)
-    #include <wolfssl/wolfcrypt/lms.h>
-    #ifdef HAVE_LIBLMS
-        #include <wolfssl/wolfcrypt/ext_lms.h>
-    #else
-        #include <wolfssl/wolfcrypt/wc_lms.h>
-    #endif
+    #include <wolfssl/wolfcrypt/wc_lms.h>
 #endif
 #if defined(WOLFSSL_HAVE_XMSS) && !defined(WOLFSSL_XMSS_VERIFY_ONLY)
-    #include <wolfssl/wolfcrypt/xmss.h>
-    #ifdef HAVE_LIBXMSS
-        #include <wolfssl/wolfcrypt/ext_xmss.h>
-    #else
-        #include <wolfssl/wolfcrypt/wc_xmss.h>
-    #endif
+    #include <wolfssl/wolfcrypt/wc_xmss.h>
 #endif
 #if defined(WOLFSSL_HAVE_SLHDSA)
     #include <wolfssl/wolfcrypt/wc_slhdsa.h>
@@ -11894,10 +11884,8 @@ static void bench_lms_sign_verify(enum wc_LmsParm parm, byte* pub)
 
 #ifndef WOLFSSL_WC_LMS_SMALL
     do {
-    #ifdef WOLFSSL_WC_LMS
         key.priv.inited = 0;
         key.state = WC_LMS_STATE_PARMSET;
-    #endif
         ret = wc_LmsKey_Reload(&key);
         if (ret) {
             printf("wc_LmsKey_Reload failed: %d\n", ret);
@@ -11918,9 +11906,6 @@ static void bench_lms_sign_verify(enum wc_LmsParm parm, byte* pub)
             printf("wc_LmsKey_GetPrivLen failed: %d\n", ret);
             goto exit_lms_sign_verify;
         }
-    #ifdef HAVE_LIBLMS
-        break;
-    #endif
     } while (bench_stats_check(start)
 #ifdef MULTI_VALUE_STATISTICS
        || runs < minimum_runs
@@ -12039,7 +12024,7 @@ void bench_lms(void)
 
 #ifndef WOLFSSL_NO_LMS_SHA256_256
 #ifdef BENCH_LMS_SLOW_KEYGEN
-#if !defined(WOLFSSL_WC_LMS) || (LMS_MAX_HEIGHT >= 15)
+#if (LMS_MAX_HEIGHT >= 15)
     bench_lms_keygen(WC_LMS_PARM_L1_H15_W2, pub);
     bench_lms_sign_verify(WC_LMS_PARM_L1_H15_W2, pub);
     bench_lms_keygen(WC_LMS_PARM_L1_H15_W4, pub);
@@ -12048,8 +12033,7 @@ void bench_lms(void)
     #define LMS_PARAMS_BENCHED
 #endif
 #endif
-#if !defined(WOLFSSL_WC_LMS) || ((LMS_MAX_LEVELS >= 2) && \
-        (LMS_MAX_HEIGHT >= 10))
+#if (LMS_MAX_LEVELS >= 2) && (LMS_MAX_HEIGHT >= 10)
     bench_lms_keygen(WC_LMS_PARM_L2_H10_W2, pub);
     bench_lms_sign_verify(WC_LMS_PARM_L2_H10_W2, pub);
     bench_lms_keygen(WC_LMS_PARM_L2_H10_W4, pub);
@@ -12061,7 +12045,7 @@ void bench_lms(void)
     bench_lms_sign_verify(WC_LMS_PARM_L2_H10_W8, pub);
 #endif
 #endif
-#if !defined(WOLFSSL_WC_LMS) || (LMS_MAX_LEVELS >= 3)
+#if (LMS_MAX_LEVELS >= 3)
     bench_lms_keygen(WC_LMS_PARM_L3_H5_W4, pub);
     bench_lms_sign_verify(WC_LMS_PARM_L3_H5_W4, pub);
     bench_lms_keygen(WC_LMS_PARM_L3_H5_W8, pub);
@@ -12069,17 +12053,16 @@ void bench_lms(void)
     #undef LMS_PARAMS_BENCHED
     #define LMS_PARAMS_BENCHED
 #endif
-#if !defined(WOLFSSL_WC_LMS) || ((LMS_MAX_LEVELS >= 3) && \
-        (LMS_MAX_HEIGHT >= 10))
+#if (LMS_MAX_LEVELS >= 3) && (LMS_MAX_HEIGHT >= 10)
     bench_lms_keygen(WC_LMS_PARM_L3_H10_W4, pub);
     bench_lms_sign_verify(WC_LMS_PARM_L3_H10_W4, pub);
 #endif
-#if !defined(WOLFSSL_WC_LMS) || (LMS_MAX_LEVELS >= 4)
+#if (LMS_MAX_LEVELS >= 4)
     bench_lms_keygen(WC_LMS_PARM_L4_H5_W8, pub);
     bench_lms_sign_verify(WC_LMS_PARM_L4_H5_W8, pub);
 #endif
 
-#if defined(WOLFSSL_WC_LMS) && !defined(LMS_PARAMS_BENCHED)
+#ifndef LMS_PARAMS_BENCHED
     bench_lms_keygen(WC_LMS_PARM_L1_H5_W1, pub);
     bench_lms_sign_verify(WC_LMS_PARM_L1_H5_W1, pub);
 #endif
@@ -12087,7 +12070,7 @@ void bench_lms(void)
 
 #ifdef WOLFSSL_LMS_SHA256_192
 #ifdef BENCH_LMS_SLOW_KEYGEN
-#if !defined(WOLFSSL_WC_LMS) || (LMS_MAX_HEIGHT >= 15)
+#if (LMS_MAX_HEIGHT >= 15)
     bench_lms_keygen(WC_LMS_PARM_SHA256_192_L1_H15_W2, pub);
     bench_lms_sign_verify(WC_LMS_PARM_SHA256_192_L1_H15_W2, pub);
     bench_lms_keygen(WC_LMS_PARM_SHA256_192_L1_H15_W4, pub);
@@ -12096,8 +12079,7 @@ void bench_lms(void)
     #define LMS_PARAMS_BENCHED
 #endif
 #endif
-#if !defined(WOLFSSL_WC_LMS) || ((LMS_MAX_LEVELS >= 2) && \
-        (LMS_MAX_HEIGHT >= 10))
+#if (LMS_MAX_LEVELS >= 2) && (LMS_MAX_HEIGHT >= 10)
     bench_lms_keygen(WC_LMS_PARM_SHA256_192_L2_H10_W2, pub);
     bench_lms_sign_verify(WC_LMS_PARM_SHA256_192_L2_H10_W2, pub);
     bench_lms_keygen(WC_LMS_PARM_SHA256_192_L2_H10_W4, pub);
@@ -12109,7 +12091,7 @@ void bench_lms(void)
     bench_lms_sign_verify(WC_LMS_PARM_SHA256_192_L2_H10_W8, pub);
 #endif
 #endif
-#if !defined(WOLFSSL_WC_LMS) || (LMS_MAX_LEVELS >= 3)
+#if (LMS_MAX_LEVELS >= 3)
     bench_lms_keygen(WC_LMS_PARM_SHA256_192_L3_H5_W4, pub);
     bench_lms_sign_verify(WC_LMS_PARM_SHA256_192_L3_H5_W4, pub);
     bench_lms_keygen(WC_LMS_PARM_SHA256_192_L3_H5_W8, pub);
@@ -12117,17 +12099,16 @@ void bench_lms(void)
     #undef LMS_PARAMS_BENCHED
     #define LMS_PARAMS_BENCHED
 #endif
-#if !defined(WOLFSSL_WC_LMS) || ((LMS_MAX_LEVELS >= 3) && \
-        (LMS_MAX_HEIGHT >= 10))
+#if (LMS_MAX_LEVELS >= 3) && (LMS_MAX_HEIGHT >= 10)
     bench_lms_keygen(WC_LMS_PARM_SHA256_192_L3_H10_W4, pub);
     bench_lms_sign_verify(WC_LMS_PARM_SHA256_192_L3_H10_W4, pub);
 #endif
-#if !defined(WOLFSSL_WC_LMS) || (LMS_MAX_LEVELS >= 4)
+#if (LMS_MAX_LEVELS >= 4)
     bench_lms_keygen(WC_LMS_PARM_SHA256_192_L4_H5_W8, pub);
     bench_lms_sign_verify(WC_LMS_PARM_SHA256_192_L4_H5_W8, pub);
 #endif
 
-#if defined(WOLFSSL_WC_LMS) && !defined(LMS_PARAMS_BENCHED)
+#ifndef LMS_PARAMS_BENCHED
     bench_lms_keygen(WC_LMS_PARM_SHA256_192_L1_H5_W1, pub);
     bench_lms_sign_verify(WC_LMS_PARM_SHA256_192_L1_H5_W1, pub);
 #endif
@@ -12208,13 +12189,6 @@ static void bench_xmss_sign_verify(const char * params)
         printf("wc_XmssKey_GetPubLen failed: %d\n", ret);
         goto exit_xmss_sign_verify;
     }
-#ifndef WOLFSSL_WC_XMSS
-    if (pkSz != XMSS_SHA256_PUBLEN) {
-        printf("error: xmss pub len: got %u, expected %d\n", pkSz,
-                XMSS_SHA256_PUBLEN);
-        goto exit_xmss_sign_verify;
-    }
-#endif
 
     ret = wc_XmssKey_GetPrivLen(&key, &skSz);
     if (ret != 0 || skSz <= 0) {
