+++ wolfssl/wolfcrypt/wc_lms.h
@@ -19,6 +19,10 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA
  */
 
+/*!
+    \file wolfssl/wolfcrypt/wc_lms.h
+ */
+
 /* Implementation based on:
  *   RFC 8554: Leighton-Micali Hash-Based Signatures
  *   https://datatracker.ietf.org/doc/html/rfc8554
@@ -58,7 +62,7 @@
  * C = Cache bits
  * To mimic the dynamic memory usage of XMSS, use 3/3.
  *
- * WOLFSSL_LMS_NO_SIGN SMOOTHING                        Default: OFF
+ * WOLFSSL_LMS_NO_SIGN_SMOOTHING                        Default: OFF
  *   Disable precalculation of next subtree.
  *   Use less dynamic memory.
  *   At certain indexes, signing will take a long time compared to the mean.
@@ -90,9 +94,9 @@
 
 #include <wolfssl/wolfcrypt/types.h>
 
-#if defined(WOLFSSL_HAVE_LMS) && defined(WOLFSSL_WC_LMS)
+#ifdef WOLFSSL_HAVE_LMS
 
-#include <wolfssl/wolfcrypt/lms.h>
+#include <wolfssl/wolfcrypt/random.h>
 #include <wolfssl/wolfcrypt/sha256.h>
 #ifdef WOLFSSL_LMS_SHAKE256
 #include <wolfssl/wolfcrypt/sha3.h>
@@ -103,6 +107,192 @@
 #if defined(WOLFSSL_NO_HASH_RAW) && !defined(WC_LMS_FULL_HASH)
     #define WC_LMS_FULL_HASH
 #endif
+
+/* Length of the Key ID. */
+#define WC_LMS_I_LEN    16
+
+/* Private key write and read callbacks. */
+typedef int (*wc_lms_write_private_key_cb)(const byte * priv, word32 privSz, void *context);
+typedef int (*wc_lms_read_private_key_cb)(byte * priv, word32 privSz, void *context);
+
+/* Return codes returned by private key callbacks. */
+enum wc_LmsRc {
+  WC_LMS_RC_NONE,
+  WC_LMS_RC_BAD_ARG,            /* Bad arg in read or write callback. */
+  WC_LMS_RC_WRITE_FAIL,         /* Write or update private key failed. */
+  WC_LMS_RC_READ_FAIL,          /* Read private key failed. */
+  WC_LMS_RC_SAVED_TO_NV_MEMORY, /* Wrote private key to nonvolatile storage. */
+  WC_LMS_RC_READ_TO_MEMORY      /* Read private key from storage. */
+};
+
+/* LMS/HSS signatures are defined by 3 parameters:
+ *   levels: number of levels of Merkle trees.
+ *   height: height of an individual Merkle tree.
+ *   winternitz: number of bits from hash used in a Winternitz chain.
+ *
+ * The acceptable parameter values are those in RFC8554:
+ *   levels = {1..8}
+ *   height = {5, 10, 15, 20, 25}
+ *   winternitz = {1, 2, 4, 8}
+ *
+ * The number of available signatures is:
+ *   N = 2 ** (levels * height)
+ *
+ * Signature sizes are determined by levels and winternitz
+ * parameters primarily, and height to a lesser extent:
+ *   - Larger levels values increase signature size significantly.
+ *   - Larger height values increase signature size moderately.
+ *   - Larger winternitz values will reduce the signature size, at
+ *     the expense of longer key generation and sign/verify times.
+ *
+ * Key generation time is strongly determined by the height of
+ * the first level tree. A 3 level, 5 height tree is much faster
+ * than 1 level, 15 height at initial key gen, even if the number
+ * of available signatures is the same.
+ * */
+
+/* Predefined LMS/HSS parameter sets for convenience.
+ *
+ * Not predefining many sets with Winternitz=1, because the signatures
+ * will be large. */
+enum wc_LmsParm {
+#ifndef WOLFSSL_NO_LMS_SHA256_256
+    WC_LMS_PARM_NONE = 0,
+    WC_LMS_PARM_L1_H5_W1 = 1,
+    WC_LMS_PARM_L1_H5_W2 = 2,
+    WC_LMS_PARM_L1_H5_W4 = 3,
+    WC_LMS_PARM_L1_H5_W8 = 4,
+    WC_LMS_PARM_L1_H10_W2 = 5,
+    WC_LMS_PARM_L1_H10_W4 = 6,
+    WC_LMS_PARM_L1_H10_W8 = 7,
+    WC_LMS_PARM_L1_H15_W2 = 8,
+    WC_LMS_PARM_L1_H15_W4 = 9,
+    WC_LMS_PARM_L1_H15_W8 = 10,
+    WC_LMS_PARM_L1_H20_W2 = 11,
+    WC_LMS_PARM_L1_H20_W4 = 12,
+    WC_LMS_PARM_L1_H20_W8 = 13,
+    WC_LMS_PARM_L2_H5_W2 = 14,
+    WC_LMS_PARM_L2_H5_W4 = 15,
+    WC_LMS_PARM_L2_H5_W8 = 16,
+    WC_LMS_PARM_L2_H10_W2 = 17,
+    WC_LMS_PARM_L2_H10_W4 = 18,
+    WC_LMS_PARM_L2_H10_W8 = 19,
+    WC_LMS_PARM_L2_H15_W2 = 20,
+    WC_LMS_PARM_L2_H15_W4 = 21,
+    WC_LMS_PARM_L2_H15_W8 = 22,
+    WC_LMS_PARM_L2_H20_W2 = 23,
+    WC_LMS_PARM_L2_H20_W4 = 24,
+    WC_LMS_PARM_L2_H20_W8 = 25,
+    WC_LMS_PARM_L3_H5_W2 = 26,
+    WC_LMS_PARM_L3_H5_W4 = 27,
+    WC_LMS_PARM_L3_H5_W8 = 28,
+    WC_LMS_PARM_L3_H10_W4 = 29,
+    WC_LMS_PARM_L3_H10_W8 = 30,
+    WC_LMS_PARM_L4_H5_W2 = 31,
+    WC_LMS_PARM_L4_H5_W4 = 32,
+    WC_LMS_PARM_L4_H5_W8 = 33,
+    WC_LMS_PARM_L4_H10_W4 = 34,
+    WC_LMS_PARM_L4_H10_W8 = 35,
+    /* H25 parameter sets for SHA-256/256 */
+    WC_LMS_PARM_L1_H25_W1 = 56,
+    WC_LMS_PARM_L1_H25_W2 = 57,
+    WC_LMS_PARM_L1_H25_W4 = 58,
+    WC_LMS_PARM_L1_H25_W8 = 59,
+    /* W1 for non-H5 heights */
+    WC_LMS_PARM_L1_H10_W1 = 60,
+    WC_LMS_PARM_L1_H15_W1 = 61,
+    WC_LMS_PARM_L1_H20_W1 = 62,
+#endif
+
+#ifdef WOLFSSL_LMS_SHA256_192
+    WC_LMS_PARM_SHA256_192_L1_H5_W1  = 36,
+    WC_LMS_PARM_SHA256_192_L1_H5_W2  = 37,
+    WC_LMS_PARM_SHA256_192_L1_H5_W4  = 38,
+    WC_LMS_PARM_SHA256_192_L1_H5_W8  = 39,
+    WC_LMS_PARM_SHA256_192_L1_H10_W2 = 40,
+    WC_LMS_PARM_SHA256_192_L1_H10_W4 = 41,
+    WC_LMS_PARM_SHA256_192_L1_H10_W8 = 42,
+    WC_LMS_PARM_SHA256_192_L1_H15_W2 = 43,
+    WC_LMS_PARM_SHA256_192_L1_H15_W4 = 44,
+    WC_LMS_PARM_SHA256_192_L1_H20_W2 = 53,
+    WC_LMS_PARM_SHA256_192_L1_H20_W4 = 54,
+    WC_LMS_PARM_SHA256_192_L1_H20_W8 = 55,
+    WC_LMS_PARM_SHA256_192_L2_H10_W2 = 45,
+    WC_LMS_PARM_SHA256_192_L2_H10_W4 = 46,
+    WC_LMS_PARM_SHA256_192_L2_H10_W8 = 47,
+    WC_LMS_PARM_SHA256_192_L3_H5_W2  = 48,
+    WC_LMS_PARM_SHA256_192_L3_H5_W4  = 49,
+    WC_LMS_PARM_SHA256_192_L3_H5_W8  = 50,
+    WC_LMS_PARM_SHA256_192_L3_H10_W4 = 51,
+    WC_LMS_PARM_SHA256_192_L4_H5_W8  = 52,
+    /* H25 for SHA-256/192 */
+    WC_LMS_PARM_SHA256_192_L1_H25_W1 = 63,
+    WC_LMS_PARM_SHA256_192_L1_H25_W2 = 64,
+    WC_LMS_PARM_SHA256_192_L1_H25_W4 = 65,
+    WC_LMS_PARM_SHA256_192_L1_H25_W8 = 66,
+    /* W1 for non-H5 heights (SHA-256/192) */
+    WC_LMS_PARM_SHA256_192_L1_H10_W1 = 67,
+    WC_LMS_PARM_SHA256_192_L1_H15_W1 = 68,
+    WC_LMS_PARM_SHA256_192_L1_H20_W1 = 69,
+    WC_LMS_PARM_SHA256_192_L1_H15_W8 = 70,
+#endif
+
+#ifdef WOLFSSL_LMS_SHAKE256
+    /* SHAKE256/256, 32-byte output */
+    WC_LMS_PARM_SHAKE_L1_H5_W1  = 100,
+    WC_LMS_PARM_SHAKE_L1_H5_W2  = 101,
+    WC_LMS_PARM_SHAKE_L1_H5_W4  = 102,
+    WC_LMS_PARM_SHAKE_L1_H5_W8  = 103,
+    WC_LMS_PARM_SHAKE_L1_H10_W1 = 104,
+    WC_LMS_PARM_SHAKE_L1_H10_W2 = 105,
+    WC_LMS_PARM_SHAKE_L1_H10_W4 = 106,
+    WC_LMS_PARM_SHAKE_L1_H10_W8 = 107,
+    WC_LMS_PARM_SHAKE_L1_H15_W1 = 108,
+    WC_LMS_PARM_SHAKE_L1_H15_W2 = 109,
+    WC_LMS_PARM_SHAKE_L1_H15_W4 = 110,
+    WC_LMS_PARM_SHAKE_L1_H15_W8 = 111,
+    WC_LMS_PARM_SHAKE_L1_H20_W1 = 112,
+    WC_LMS_PARM_SHAKE_L1_H20_W2 = 113,
+    WC_LMS_PARM_SHAKE_L1_H20_W4 = 114,
+    WC_LMS_PARM_SHAKE_L1_H20_W8 = 115,
+    WC_LMS_PARM_SHAKE_L1_H25_W1 = 116,
+    WC_LMS_PARM_SHAKE_L1_H25_W2 = 117,
+    WC_LMS_PARM_SHAKE_L1_H25_W4 = 118,
+    WC_LMS_PARM_SHAKE_L1_H25_W8 = 119,
+    /* SHAKE256/192, 24-byte output */
+    WC_LMS_PARM_SHAKE192_L1_H5_W1  = 120,
+    WC_LMS_PARM_SHAKE192_L1_H5_W2  = 121,
+    WC_LMS_PARM_SHAKE192_L1_H5_W4  = 122,
+    WC_LMS_PARM_SHAKE192_L1_H5_W8  = 123,
+    WC_LMS_PARM_SHAKE192_L1_H10_W1 = 124,
+    WC_LMS_PARM_SHAKE192_L1_H10_W2 = 125,
+    WC_LMS_PARM_SHAKE192_L1_H10_W4 = 126,
+    WC_LMS_PARM_SHAKE192_L1_H10_W8 = 127,
+    WC_LMS_PARM_SHAKE192_L1_H15_W1 = 128,
+    WC_LMS_PARM_SHAKE192_L1_H15_W2 = 129,
+    WC_LMS_PARM_SHAKE192_L1_H15_W4 = 130,
+    WC_LMS_PARM_SHAKE192_L1_H15_W8 = 131,
+    WC_LMS_PARM_SHAKE192_L1_H20_W1 = 132,
+    WC_LMS_PARM_SHAKE192_L1_H20_W2 = 133,
+    WC_LMS_PARM_SHAKE192_L1_H20_W4 = 134,
+    WC_LMS_PARM_SHAKE192_L1_H20_W8 = 135,
+    WC_LMS_PARM_SHAKE192_L1_H25_W1 = 136,
+    WC_LMS_PARM_SHAKE192_L1_H25_W2 = 137,
+    WC_LMS_PARM_SHAKE192_L1_H25_W4 = 138,
+    WC_LMS_PARM_SHAKE192_L1_H25_W8 = 139,
+#endif
+};
+
+/* enum wc_LmsState is to help track the state of an LMS/HSS Key. */
+enum wc_LmsState {
+    WC_LMS_STATE_FREED,      /* Key has been freed from memory. */
+    WC_LMS_STATE_INITED,     /* Key has been inited, ready to set params.*/
+    WC_LMS_STATE_PARMSET,    /* Params are set, ready to MakeKey or Reload. */
+    WC_LMS_STATE_OK,         /* Able to sign signatures and verify. */
+    WC_LMS_STATE_VERIFYONLY, /* A public only LmsKey. */
+    WC_LMS_STATE_BAD,        /* Can't guarantee key's state. */
+    WC_LMS_STATE_NOSIGS      /* Signatures exhausted. */
+};
 
 #ifdef WOLFSSL_LMS_MAX_LEVELS
     /* Maximum number of levels of trees supported by implementation. */
@@ -546,7 +736,7 @@ typedef struct HssPrivKey {
 #endif
 } HssPrivKey;
 
-struct LmsKey {
+typedef struct LmsKey {
     /* Public key. */
     ALIGN16 byte pub[HSS_PUBLIC_KEY_LEN(LMS_MAX_NODE_LEN)];
 #ifndef WOLFSSL_LMS_VERIFY_ONLY
@@ -575,7 +765,48 @@ struct LmsKey {
     /* Device Identifier. */
     int devId;
 #endif
-};
+} LmsKey;
+
+#ifdef __cplusplus
+    extern "C" {
+#endif
+
+WOLFSSL_API int  wc_LmsKey_Init(LmsKey * key, void * heap, int devId);
+WOLFSSL_API int  wc_LmsKey_SetLmsParm(LmsKey * key, enum wc_LmsParm lmsParm);
+WOLFSSL_API int  wc_LmsKey_SetParameters(LmsKey * key, int levels,
+    int height, int winternitz);
+WOLFSSL_API int  wc_LmsKey_GetParameters(const LmsKey * key, int * levels,
+    int * height, int * winternitz);
+#ifndef WOLFSSL_LMS_VERIFY_ONLY
+WOLFSSL_API int  wc_LmsKey_SetWriteCb(LmsKey * key,
+    wc_lms_write_private_key_cb write_cb);
+WOLFSSL_API int  wc_LmsKey_SetReadCb(LmsKey * key,
+    wc_lms_read_private_key_cb read_cb);
+WOLFSSL_API int  wc_LmsKey_SetContext(LmsKey * key, void * context);
+WOLFSSL_API int  wc_LmsKey_MakeKey(LmsKey * key, WC_RNG * rng);
+WOLFSSL_API int  wc_LmsKey_Reload(LmsKey * key);
+WOLFSSL_API int  wc_LmsKey_GetPrivLen(const LmsKey * key, word32 * len);
+WOLFSSL_API int  wc_LmsKey_Sign(LmsKey * key, byte * sig, word32 * sigSz,
+    const byte * msg, int msgSz);
+WOLFSSL_API int  wc_LmsKey_SigsLeft(LmsKey * key);
+#endif /* ifndef WOLFSSL_LMS_VERIFY_ONLY */
+WOLFSSL_API void wc_LmsKey_Free(LmsKey * key);
+WOLFSSL_API int  wc_LmsKey_GetSigLen(const LmsKey * key, word32 * len);
+WOLFSSL_API int  wc_LmsKey_GetPubLen(const LmsKey * key, word32 * len);
+WOLFSSL_API int  wc_LmsKey_ExportPub(LmsKey * keyDst, const LmsKey * keySrc);
+WOLFSSL_API int  wc_LmsKey_ExportPubRaw(const LmsKey * key, byte * out,
+    word32 * outLen);
+WOLFSSL_API int  wc_LmsKey_ImportPubRaw(LmsKey * key, const byte * in,
+    word32 inLen);
+WOLFSSL_API int  wc_LmsKey_Verify(LmsKey * key, const byte * sig, word32 sigSz,
+    const byte * msg, int msgSz);
+WOLFSSL_API const char * wc_LmsKey_ParmToStr(enum wc_LmsParm lmsParm);
+WOLFSSL_API const char * wc_LmsKey_RcToStr(enum wc_LmsRc lmsRc);
+
+WOLFSSL_API int wc_LmsKey_GetKid(LmsKey * key, const byte ** kid,
+    word32* kidSz);
+WOLFSSL_API const byte * wc_LmsKey_GetKidFromPrivRaw(const byte * priv,
+    word32 privSz);
 
 int wc_hss_make_key(LmsState* state, WC_RNG* rng, byte* priv_raw,
     HssPrivKey* priv_key, byte* priv_data, byte* pub);
@@ -588,6 +819,10 @@ WOLFSSL_API
 int wc_hss_verify(LmsState* state, const byte* pub, const byte* msg,
     word32 msgSz, const byte* sig, word32 sigSz);
 
-#endif /* WOLFSSL_HAVE_LMS && WOLFSSL_WC_LMS */
+#ifdef __cplusplus
+} /* extern "C" */
+#endif
+
+#endif /* WOLFSSL_HAVE_LMS */
 
 #endif /* WC_LMS_H */
